+++ examples/ai-core/src/embed-many/azure.ts
@@ -4,7 +4,7 @@ import 'dotenv/config';
 
 async function main() {
   const { embeddings, usage } = await embedMany({
-    model: azure.embedding('my-embedding-deployment'),
+    model: azure.embedding('text-embedding-3-large'), // use your own deployment
     values: [
       'sunny day at the beach',
       'rainy afternoon in the city',