+++ packages/ai/src/generate-text/output.ts
@@ -1,18 +1,16 @@
 import { LanguageModelV3CallOptions } from '@ai-sdk/provider';
 import {
   asSchema,
+  FlexibleSchema,
   safeParseJSON,
   safeValidateTypes,
-  Schema,
 } from '@ai-sdk/provider-utils';
-import * as z3 from 'zod/v3';
-import * as z4 from 'zod/v4';
 import { NoObjectGeneratedError } from '../error/no-object-generated-error';
-import { DeepPartial } from '../util/deep-partial';
-import { parsePartialJson } from '../util/parse-partial-json';
 import { FinishReason } from '../types/language-model';
 import { LanguageModelResponseMetadata } from '../types/language-model-response-metadata';
 import { LanguageModelUsage } from '../types/usage';
+import { DeepPartial } from '../util/deep-partial';
+import { parsePartialJson } from '../util/parse-partial-json';
 
 export interface Output<OUTPUT, PARTIAL> {
   readonly type: 'object' | 'text';
@@ -50,10 +48,7 @@ export const text = (): Output<string, string> => ({
 export const object = <OUTPUT>({
   schema: inputSchema,
 }: {
-  schema:
-    | z4.core.$ZodType<OUTPUT, any>
-    | z3.Schema<OUTPUT, z3.ZodTypeDef, any>
-    | Schema<OUTPUT>;
+  schema: FlexibleSchema<OUTPUT>;
 }): Output<OUTPUT, DeepPartial<OUTPUT>> => {
   const schema = asSchema(inputSchema);